FixSuggester/requirements.txt
fastapi
uvicorn
redis
asyncpg
fastapi
uvicorn
redis
asyncpg
google-genai
pydantic-settings
python-dotenv
pika
